requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
python_jose>=3.3.0
passlib>=1.7.4
pydantic[email]
python-multipart>=0.0.20
jinja2>=3.1.5
pyJWT>=2.10.1
itsdangerous>=2.2.0